backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.10.4
pydantic-settings==2.6.1
SQLAlchemy==2.0.39
psycopg[binary]==3.2.4
python-multipart==0.0.9
moviepy==1.0.3
Pillow==11.0.0
numpy==2.1.1
imageio-ffmpeg==0.4.9
aiofiles==23.2.1
requests==2.32.3